+++ app/src/components/EventCard.svelte
